Move pose processing from GamePlayDirector
