+++ Assets/Project/Scripts/GamePlayScene/GamePausePopup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f0ae0d7b2bd214025a6e5c1080d57690
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
